Changes (Read Description)
LOTS of changes to code to add the health bars and finalize the level screen. Most of the sprites that will be used on the overworld screen will be the ones provided in the tutorial due to time constraints. However, if I can find time and suitable sprites, I will replace them.